Struct tdigest_ch::TDigest

pub struct TDigest { /* private fields */ }

T-digest data structure for approximating the quantiles of a distribution.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();

// Add some elements.
digest.insert(1.0);
digest.insert(2.0);
digest.insert(3.0);

// Get the median of the distribution.
let quantile = digest.quantile(0.5);
assert_eq!(quantile, 2.0);

Implementations

impl TDigest

pub fn new() -> Self

Creates an empty TDigest.

Examples

use tdigest_ch::TDigest;
let digest = TDigest::new();

pub fn builder() -> TDigestBuilder

Creates a TDigestBuilder to configure a TDigest.

This is the same as TDigestBuilder::new().

pub fn append(&mut self, other: &mut TDigest)

Moves all the elements of other into self, leaving other empty.

Examples

use tdigest_ch::TDigest;

let mut a = TDigest::from([-10.0, 1.0, 2.0, 2.0, 3.0]);
let mut b = TDigest::from([-20.0, 5.0, 43.0]);

a.append(&mut b);

assert_eq!(a.len(), 8);
assert!(b.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the t-digest.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();
assert_eq!(digest.len(), 0);
digest.insert(1.0);
assert_eq!(digest.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the t-digest contains no elements.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();
assert!(digest.is_empty());
digest.insert(1.0);
assert!(!digest.is_empty());

pub fn clear(&mut self)

Clears the t-digest, removing all values.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();
digest.insert(1.0);
digest.clear();
assert!(digest.is_empty());

pub fn quantile(&mut self, level: f64) -> f32

Returns the estimated quantile of the t-digest.

This method expects self to be mutable, since the t-digest may be compressed. If you require an immutable, shared reference to compute quantiles, consider using quantiles instead.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::from([1.0, 2.0, 3.0, 4.0, 5.0]);
assert_eq!(digest.quantile(0.0), 1.0);
assert_eq!(digest.quantile(0.5), 3.0);
assert_eq!(digest.quantile(1.0), 5.0);

pub fn quantiles(&mut self) -> Quantiles<'_>

Creates an immutable quantile estimator from the t-digest.

Examples

use std::thread;

use tdigest_ch::TDigest;

let mut digest = TDigest::from([1.0, 2.0, 3.0, 4.0, 5.0]);
let quantiles = digest.quantiles();

thread::scope(|s| {
    s.spawn(|| {
        assert_eq!(quantiles.get(0.0), 1.0);
    });
    s.spawn(|| {
        assert_eq!(quantiles.get(0.5), 3.0);
    });
    s.spawn(|| {
        assert_eq!(quantiles.get(1.0), 5.0);
    });
});

pub fn insert(&mut self, value: f32)

Adds a value to the t-digest.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();

digest.insert(1.0);
digest.insert(2.0);
assert_eq!(digest.len(), 2);

pub fn insert_many(&mut self, value: f32, count: usize)

Adds multiple values to the t-digest.

Examples

use tdigest_ch::TDigest;

let mut digest = TDigest::new();

digest.insert_many(1.0, 1);
digest.insert_many(2.0, 2);
assert_eq!(digest.len(), 3);